abstracted VideoExample
