+++ examples/components/FaceDeformationVideoExample.js
@@ -1,65 +1,52 @@
 import React from 'react';
-import classNames from 'classnames';
 import Dat, { DatNumber, DatBoolean } from 'react-dat-gui';
 import '!style!css!react-dat-gui/lib/Dat.css';
+import _ from 'lodash';
 
-import RaisedButton from 'material-ui/RaisedButton';
+import VideoExample from './VideoExample';
 
 import Tracker from 'clmtrackr/js/Tracker';
-import {
-  supportsVideo,
-  supportsH264BaselineVideo,
-  supportsOggTheoraVideo,
-  supportsUserMedia,
-  loadVideo
-} from 'clmtrackr/js/utils/video';
-import { requestAnimFrame } from 'clmtrackr/js/utils/anim';
-
 import { ParamHolder } from 'clmtrackr/js/deformers/presets';
-
-import TrackerContainer from 'clmtrackr/ui/container/TrackerContainer';
 import Deformer from 'clmtrackr/js/deformers/twgl';
 
 import './FaceDeformationVideoExample.styl';
 
 
-const MEDIA_SIZE = { width: 370, height: 288 };
-
-
-export default class FaceDeformationVideoExample extends React.Component {
+export default class FaceDeformationVideoExample extends VideoExample {
   constructor () {
     super();
-    this.state = {
-      tracker: null,
+    this.state = _.merge(this.state, {
       deformer: null,
-
       data: new ParamHolder(),
-      startButtonDisabled: true,
-      showScore: false,
-
       positions: null,
-      hideMedia: false
-    };
+      title: 'Face deformation'
+    });
 
-    this._boundOnFrame = this._onFrame.bind(this);
+    this.mediaSize = { width: 370, height: 288 };
+    this.oggVideoSrc = 'media/franck.ogv';
+    this.mp4VideoSrc = null;
+  }
+
+  newTracker () {
+    return new Tracker({ stopOnConvergence: true });
   }
 
   componentDidMount () {
-    const tracker = new Tracker({ stopOnConvergence: true });
-    tracker.init();
-    tracker.convergenceThreshold = 1.8;
+    super.componentDidMount();
 
     const deformer = new Deformer();
     const trackerContainer = this.refs.trackerContainer;
     deformer.init(trackerContainer.refs.deformerCanvas);
 
-    this.setState({ tracker, deformer });
-
-    tracker.on('converged', () => {
-      this._setupFaceDeformation();
-    });
+    this.setState({ deformer });
 
     setTimeout(() => {
+      const tracker = this.state.tracker;
+      tracker.on('converged', () => {
+        this._setupFaceDeformation();
+      });
+      tracker.convergenceThreshold = 1.8;
+
       this._setupVideoStream();
     });
   }
@@ -117,79 +104,8 @@ export default class FaceDeformationVideoExample extends React.Component {
     }
   }
 
-  _setupVideoStream () {
-    // check for camerasupport
-    if (supportsUserMedia()) {
-      // set up stream
-      loadVideo((err, stream) => {
-        if (err) {
-          this._insertAltVideo();
-          this.setState({ showAltInstructions: true });
-          alert('There was some problem trying to fetch video from your webcam, using a fallback video instead.');
-          return;
-        }
-
-        this.setState({ mediaSrc: stream });
-      });
-    } else {
-      this._insertAltVideo();
-      this.setState({ showAltInstructions: true });
-      alert('Your browser does not seem to support getUserMedia, using a fallback video instead.');
-    }
-
-    const trackerContainer = this.refs.trackerContainer;
-    const video = trackerContainer.refs.media;
-    video.addEventListener('canplay', () => {
-      this.setState({ startButtonDisabled: false });
-    });
-  }
-
-  _onFrame () {
-    // Update overlay
-    const trackerContainer = this.refs.trackerContainer;
-    if (trackerContainer) {
-      // TODO: This should really draw to the same canvas as the deformer,
-      // Need to add webgl support to `tracker#draw`
-      const tracker = this.state.tracker;
-      const cc = trackerContainer.refs.canvas.getContext('2d');
-      cc.clearRect(0, 0, MEDIA_SIZE.width, MEDIA_SIZE.height);
-      tracker.draw(cc.canvas);
-    }
-    requestAnimFrame(this._boundOnFrame);
-  }
-
-  _resetTracker () {
-    const tracker = this.state.tracker;
-    const trackerContainer = this.refs.trackerContainer;
-    const overlayCC = trackerContainer.refs.canvas.getContext('2d');
-    overlayCC.clearRect(0, 0, MEDIA_SIZE.width, MEDIA_SIZE.height);
-    tracker.stop();
-    tracker.reset();
-  }
-
-  _insertAltVideo () {
-    if (supportsVideo()) {
-      if (supportsOggTheoraVideo()) {
-        this.setState({ mediaSrc: 'media/franck.ogv' });
-      } else if (supportsH264BaselineVideo()) {
-        alert('no mp4 video');
-      } else {
-        return false;
-      }
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  _startVideo () {
-    // start video
-    const trackerContainer = this.refs.trackerContainer;
-    const video = trackerContainer.refs.media;
-    // start tracking
-    this._resetTracker();
-    this.state.tracker.start(video);
-    this.setState({ showScore: true });
+  startVideo () {
+    super.startVideo();
     // start loop to draw face
     this._onFrame();
   }
@@ -197,6 +113,27 @@ export default class FaceDeformationVideoExample extends React.Component {
   _handleDatUpdate (dataUpdated) {
     this.setState({ data: dataUpdated });
     setTimeout(() => this._drawDeformedFace());
+  }
+
+  renderInstructions () {
+    return (
+      <ol>
+        <li>allow the page to your use webcamera</li>
+        <li>make sure that your face is clearly visible in the video, and click start</li>
+        <li>keep your head still until the model fits your face properly</li>
+        <li>choose a preset or fiddle with the parameters on the right hand side, to deform the captured face</li>
+      </ol>
+    );
+  }
+
+  renderAltInstructions () {
+    return (
+      <ol>
+        <li>click start</li>
+        <li>wait till model fits the face and a mask is applied</li>
+        <li>try out different masks from the dropdown</li>
+      </ol>
+    );
   }
 
   render () {
@@ -218,78 +155,18 @@ export default class FaceDeformationVideoExample extends React.Component {
     datControls.push(<DatBoolean path='drawGrid' label='Draw Grid' key={50} />);
     datControls.push(<DatBoolean path='drawFace' label='Draw Face' key={51} />);
 
-    let instructions;
-    if (!this.state.showAltInstructions) {
-      instructions = (
-        <div>
-          <p>To try it out:</p>
-          <ol>
-            <li>allow the page to your use webcamera</li>
-            <li>make sure that your face is clearly visible in the video, and click start</li>
-            <li>keep your head still until the model fits your face properly</li>
-            <li>choose a preset or fiddle with the parameters on the right hand side, to deform the captured face</li>
-          </ol>
-        </div>
-      );
-    } else {
-      instructions = (
-        <div>
-          <p>There was some problem trying to capture your webcamera, please check that your browser supports WebRTC. Instead we'll use a static image. To try it out:</p>
-          <ol>
-            <li>click start</li>
-            <li>wait till the model fits the face</li>
-            <li>choose a preset or fiddle with the parameters on the right hand side, to deform the captured face</li>
-          </ol>
-        </div>
-      );
-    }
-
-    const startButtonDisabled = this.state.startButtonDisabled;
-    let scoreText;
-    if (this.state.showScore) {
-      scoreText = <p>Please keep head still while model fits</p>;
-    }
-
-    return (
-      <div
-        className={classNames('face-deformation-video-example-cmpt', {
-          'hide-media': this.state.hideMedia
-        })}
+    const children = (
+      <Dat
+        data={this.state.data}
+        onUpdate={this._handleDatUpdate.bind(this)}
       >
-        <h1>Face deformation</h1>
-
-        <TrackerContainer
-          ref='trackerContainer'
-          mediaType={'video'}
-          mediaSrc={this.state.mediaSrc}
-          mediaSize={MEDIA_SIZE}
-          showStats={true}
-          tracker={this.state.tracker}
-          hasDeformer={true}
-        />
-
-        <div>
-          <RaisedButton
-            label={startButtonDisabled ? 'wait, loading video' : 'start'}
-            disabled={startButtonDisabled}
-            onClick={this._startVideo.bind(this)}
-          />
-        </div>
-        {scoreText}
-        <div>
-          <p>This is an example of face deformation using the javascript library <a href='https://github.com/auduno/clmtrackr'><em>clmtrackr</em></a>.</p>
-          <p>Note that this example needs support for WebGL, and works best in Google Chrome.</p>
-
-          {instructions}
-        </div>
-
-        <Dat
-          data={this.state.data}
-          onUpdate={this._handleDatUpdate.bind(this)}
-        >
-          {datControls}
-        </Dat>
-      </div>
+        {datControls}
+      </Dat>
     );
+
+    return super.render({
+      children,
+      className: 'face-deformation-video-example-cmpt'
+    });
   }
 }
