+++ examples/components/FaceDeformationVideoExample.styl
@@ -2,8 +2,3 @@
   .react-dat-gui
     bottom 0
     top auto
-
-  &.hide-media
-    .tracker-container-cmpt
-      .media
-        visibility hidden
