+++ examples/components/FaceMaskExample.styl
@@ -1,9 +1,4 @@
 .face-mask-example-cmpt
-  .control-row
+  .mask-select
     display flex
-    justify-content space-between
     align-items center
-
-    .mask-select
-      display flex
-      align-items center
