+++ examples/components/VideoExample.styl
@@ -1,0 +1,10 @@
+.video-example-cmpt
+  .control-row
+    display flex
+    justify-content space-between
+    align-items center
+
+  &.hide-media
+    .tracker-container-cmpt
+      .media
+        visibility hidden
